优化了接口文档，引入了redis缓存

+++ IT-server/src/main/java/com/xiaoyan/service/ArticlesService.java
@@ -9,7 +9,7 @@ import java.util.List;
 @Service
 public interface ArticlesService {
 
-    Integer getCount();
+    Long getCount();
 
     void upload(ArticleDTO articleDTO);
